PINCH AREA: solving the drag
Basically what we need to do is to not anchors.fill the Canvas element to its parent. Instead, we will set its dimension using the width and height properties. In this case I am setting exactly the same dimension as the root. Be aware that if you only comment the anchors.fill line but don't define a dimension, you will not be able to see the Canvas element because, by default, a Qt QML element has dimension 0x0 pixels. (I wonder why...)