Increasing the Canvas area
What I am doing now is creating a larger Canvas area so, when you drag it you still have canvas space to work on. So I created a canvasScale factor to be used as the multiplying factor for the canvas area. In this example I am increasing 5 times, but you can feel free to try it. Be aware that there is a size limit that the canvas can achieve (in my machine a canvas of size 100000 x 100000 does not work).

To make sure that the canvas will begin centralized in the application window, I changed the x and y position accordingly. No matter how big it is, your canvas will always be centralized. One might ask: why not use the anchors.centerIn? Because if you do so, you won't be able to drag the canvas anymore.

Apart from that, I just took out the Button from the Canvas so we can have it fixed no matter how you manipulate the canvas through the pinch gesture. You will notice that I could not use the anchors as well and I had to position the Button using the x and y properties. As far as I know, the anchors use as reference the elements identification. But for some reason I cannot use root.bottom as a reference. I could have created an outside rectangle tho and put everything else inside it, but I choose to show you how to do it using the x and y properties instead. In a more complex development one should prefer using the anchors over the position parameters because it allows you to create a better control between your elements.

I hope you enjoyed it!
Cheers!